break query into its own file
